+++ src/Zns.ts
@@ -6,7 +6,7 @@ import _ from 'lodash';
 import { ResolutionResult, SourceDefinition } from './types';
 import NamingService from './NamingService';
 
-const DefaultSource = 'https://api.zilliqa.com/';
+const DefaultSource = 'https://api.zilliqa.com';
 const NullAddress = '0x0000000000000000000000000000000000000000';
 
 const NetworkIdMap = {
@@ -20,24 +20,24 @@ const RegistryMap = {
 };
 
 const UrlMap = {
-  mainnet: 'https://api.zilliqa.com/',
-  testnet: 'https://dev-api.zilliqa.com/',
-  localnet: 'http://localhost:4201/',
+  mainnet: 'https://api.zilliqa.com',
+  testnet: 'https://dev-api.zilliqa.com',
+  localnet: 'http://localhost:4201',
 };
 
 const UrlNetworkMap = (url: string) => {
   const invert = _(UrlMap)
     .invert()
     .value();
-  return url.endsWith('/') ? invert[url] : invert[url + '/'];
+  return invert[url];
 };
 
 export default class Zns extends NamingService {
   readonly network: string;
   readonly url: string;
-  private registryAddress: string;
-  registry: Contract;
-  zilliqa: Zilliqa;
+  readonly registryAddress?: string;
+  private registry?: Contract;
+  private zilliqa: Zilliqa;
 
   constructor(source: string | boolean | SourceDefinition = true) {
     super();
@@ -51,9 +51,15 @@ export default class Zns extends NamingService {
     if (!this.url) {
       throw new Error('Unspecified url in Namicorn ZNS configuration');
     }
-    this.registryAddress = RegistryMap[this.network];
-    if (this.registryAddress)
+    this.registryAddress = source.registry
+      ? source.registry
+      : RegistryMap[this.network];
+    if (this.registryAddress) {
+      this.registryAddress = this.registryAddress.startsWith('0x')
+        ? toBech32Address(this.registryAddress)
+        : this.registryAddress;
       this.registry = this.zilliqa.contracts.at(this.registryAddress);
+    }
   }
 
   async getContractField(
@@ -61,9 +67,9 @@ export default class Zns extends NamingService {
     field: string,
     keys: string[] = [],
   ): Promise<any> {
-    let result = await contract.getSubState(
-      field, keys.map(k => JSON.stringify(k))
-    ) || {}
+    let result =
+      (await contract.getSubState(field, keys.map(k => JSON.stringify(k)))) ||
+      {};
     return result[field];
   }
 
@@ -133,6 +139,7 @@ export default class Zns extends NamingService {
   isSupportedNetwork(): boolean {
     return this.registryAddress != null;
   }
+
   protected normalizeSource(
     source: string | boolean | SourceDefinition,
   ): SourceDefinition {
@@ -141,12 +148,8 @@ export default class Zns extends NamingService {
         return { url: DefaultSource, network: 'mainnet' };
       }
       case 'string': {
-        const formatedSource = !source.endsWith('/')
-          ? source + '/'
-          : (source as string);
-
         return {
-          url: formatedSource,
+          url: source as string,
           network: UrlNetworkMap(source),
         };
       }
@@ -155,16 +158,18 @@ export default class Zns extends NamingService {
         if (typeof source.network == 'number') {
           source.network = NetworkIdMap[source.network];
         }
+        if (source.registry) {
+          source.network = source.network ? source.network : 'mainnet';
+          source.url = source.url ? source.url : DefaultSource;
+        }
         if (source.network && !source.url) {
           source.url = UrlMap[source.network];
         }
         if (source.url && !source.network) {
-          source.url = source.url.endsWith('/') ? source.url : source.url + '/';
           source.network = UrlNetworkMap(source.url);
         }
         return source;
       }
     }
   }
-
 }